added some trade endpoints
