+++ package.json
@@ -13,7 +13,6 @@
     "@emotion/styled": "^11.8.1",
     "@heroicons/react": "^1.0.6",
     "@types/crypto-js": "^4.1.1",
-    "@types/lodash": "^4.14.191",
     "@types/node": "^14.14.31",
     "@types/react": "^17.0.2",
     "axios": "^0.24.0",
@@ -23,12 +22,10 @@
     "crypto-js": "^4.1.1",
     "date-fns": "^2.28.0",
     "date-fns-tz": "^1.3.3",
-    "dayjs": "^1.11.7",
     "dotenv": "^8.2.0",
     "formik": "^2.2.9",
     "http-status-codes": "^2.1.4",
     "imagemin-svgo": "^9.0.0",
-    "lodash": "^4.17.21",
     "next": "^12.1.0",
     "next-compose-plugins": "^2.2.1",
     "next-optimized-images": "^2.6.2",
